Code Review Update for #671
