+++ fhir-operation-bulkdata/src/main/java/com/ibm/fhir/operation/bulkdata/model/PollingLocationResponse.java
@@ -214,11 +214,8 @@ public class PollingLocationResponse {
                         }
                         generator.writeEnd();
                     }
-
                     generator.writeEnd();
-
                 }
-
                 o = writer.toString();
             }
             return o;
